Cleaning the folders
