--- cmap.py
@@ -1,1 +1,0 @@
-cmap =[(  0,   0, 255),(  0,   2, 253),(  0,   4, 251),(  0,   6, 249),(  0,   8, 247),(  0,  10, 245),(  0,  12, 243),(  0,  14, 241),(  0,  16, 239),(  0,  18, 237),(  0,  20, 235),(  0,  22, 233),(  0,  24, 231),(  0,  26, 229),(  0,  28, 227),(  0,  30, 225),(  0,  32, 223),(  0,  34, 221),(  0,  36, 219),(  0,  38, 217),(  0,  40, 215),(  0,  42, 213),(  0,  44, 211),(  0,  46, 209),(  0,  48, 207),(  0,  50, 205),(  0,  52, 203),(  0,  54, 201),(  0,  56, 199),(  0,  58, 197),(  0,  60, 195),(  0,  62, 193),(  0,  64, 191),(  0,  66, 189),(  0,  68, 187),(  0,  70, 185),(  0,  72, 183),(  0,  74, 181),(  0,  76, 179),(  0,  78, 177),(  0,  80, 175),(  0,  82, 173),(  0,  84, 171),(  0,  86, 169),(  0,  88, 167),(  0,  90, 165),(  0,  92, 163),(  0,  94, 161),(  0,  96, 159),(  0,  98, 157),(  0, 100, 155),(  0, 102, 153),(  0, 104, 151),(  0, 106, 149),(  0, 108, 147),(  0, 110, 145),(  0, 112, 143),(  0, 114, 141),(  0, 116, 139),(  0, 118, 137),(  0, 120, 135),(  0, 122, 133),(  0, 124, 131),(  0, 126, 129),(  0, 128, 127),(  0, 130, 125),(  0, 132, 123),(  0, 134, 121),(  0, 136, 119),(  0, 138, 117),(  0, 140, 115),(  0, 142, 113),(  0, 144, 111),(  0, 146, 109),(  0, 148, 107),(  0, 150, 105),(  0, 152, 103),(  0, 154, 101),(  0, 156,  99),(  0, 158,  97),(  0, 160,  95),(  0, 162,  93),(  0, 164,  91),(  0, 166,  89),(  0, 168,  87),(  0, 170,  85),(  0, 172,  83),(  0, 174,  81),(  0, 176,  79),(  0, 178,  77),(  0, 180,  75),(  0, 182,  73),(  0, 184,  71),(  0, 186,  69),(  0, 188,  67),(  0, 190,  65),(  0, 192,  63),(  0, 194,  61),(  0, 196,  59),(  0, 198,  57),(  0, 200,  55),(  0, 202,  53),(  0, 204,  51),(  0, 206,  49),(  0, 208,  47),(  0, 210,  45),(  0, 212,  43),(  0, 214,  41),(  0, 216,  39),(  0, 218,  37),(  0, 220,  35),(  0, 222,  33),(  0, 224,  31),(  0, 226,  29),(  0, 228,  27),(  0, 230,  25),(  0, 232,  23),(  0, 234,  21),(  0, 236,  19),(  0, 238,  17),(  0, 240,  15),(  0, 242,  13),(  0, 244,  11),(  0, 246,   9),(  0, 248,   7),(  0, 250,   5),(  0, 252,   3),(  0, 254,   1),(  0, 254,   0),(  2, 252,   0),(  4, 250,   0),(  6, 248,   0),(  9, 245,   0),( 11, 243,   0),( 13, 242,   0),( 15, 240,   0),( 16, 238,   0),( 18, 236,   0),( 20, 234,   0),( 22, 232,   0),( 25, 229,   0),( 27, 227,   0),( 29, 226,   0),( 31, 224,   0),( 32, 222,   0),( 34, 220,   0),( 36, 218,   0),( 38, 216,   0),( 41, 213,   0),( 43, 212,   0),( 45, 210,   0),( 47, 208,   0),( 48, 206,   0),( 50, 204,   0),( 52, 202,   0),( 54, 200,   0),( 57, 197,   0),( 59, 196,   0),( 61, 194,   0),( 63, 192,   0),( 65, 190,   0),( 66, 188,   0),( 68, 186,   0),( 70, 184,   0),( 73, 181,   0),( 75, 180,   0),( 77, 178,   0),( 79, 176,   0),( 81, 174,   0),( 82, 172,   0),( 84, 170,   0),( 86, 168,   0),( 89, 165,   0),( 91, 164,   0),( 93, 162,   0),( 95, 160,   0),( 97, 158,   0),( 98, 156,   0),(100, 154,   0),(102, 152,   0),(105, 149,   0),(107, 148,   0),(109, 146,   0),(111, 144,   0),(113, 142,   0),(114, 140,   0),(116, 138,   0),(118, 136,   0),(121, 133,   0),(123, 132,   0),(125, 130,   0),(127, 128,   0),(129, 126,   0),(131, 124,   0),(132, 122,   0),(134, 120,   0),(137, 117,   0),(139, 115,   0),(141, 114,   0),(143, 112,   0),(145, 110,   0),(147, 108,   0),(148, 106,   0),(150, 104,   0),(153, 101,   0),(155,  99,   0),(157,  98,   0),(159,  96,   0),(161,  94,   0),(163,  92,   0),(164,  90,   0),(166,  88,   0),(169,  85,   0),(171,  83,   0),(173,  82,   0),(175,  80,   0),(177,  78,   0),(179,  76,   0),(180,  74,   0),(182,  72,   0),(185,  69,   0),(187,  67,   0),(189,  66,   0),(191,  64,   0),(193,  62,   0),(195,  60,   0),(196,  58,   0),(198,  56,   0),(201,  53,   0),(203,  51,   0),(205,  50,   0),(207,  48,   0),(209,  46,   0),(211,  44,   0),(212,  42,   0),(214,  40,   0),(217,  37,   0),(219,  35,   0),(221,  34,   0),(223,  32,   0),(225,  30,   0),(227,  28,   0),(228,  26,   0),(230,  24,   0),(233,  21,   0),(235,  19,   0),(237,  18,   0),(239,  16,   0),(241,  14,   0),(243,  12,   0),(244,  10,   0),(246,   8,   0),(249,   5,   0),(251,   4,   0),(253,   2,   0),(255,   0,   0)]
