--- colormap.py
@@ -1,40 +1,0 @@
-import sys
-
-EPSILON = sys.float_info.epsilon  # Smallest possible difference.
-
-
-def convert_to_rgb(minval, maxval, val, colors):
-    # `colors` is a series of RGB colors delineating a series of
-    # adjacent linear color gradients between each pair.
-
-    # Determine where the given value falls proportionality within
-    # the range from minval->maxval and scale that fractional value
-    # by the total number in the `colors` palette.
-    i_f = float(val - minval) / float(maxval - minval) * (len(colors) - 1)
-
-    # Determine the lower index of the pair of color indices this
-    # value corresponds and its fractional distance between the lower
-    # and the upper colors.
-    i, f = int(i_f // 1), i_f % 1  # Split into whole & fractional parts.
-
-    # Does it fall exactly on one of the color points?
-    if f < EPSILON:
-        return colors[i]
-    else:  # Return a color linearly interpolated in the range between it and
-        # the following one.
-        (r1, g1, b1), (r2, g2, b2) = colors[i], colors[i + 1]
-        return int(r1 + f * (r2 - r1)), int(g1 + f * (g2 - g1)), int(b1 + f * (b2 - b1))
-
-
-if __name__ == "__main__":
-    minval, maxval = 0, 255
-    steps = 255
-    delta = float(maxval - minval) / steps
-    colors = [(0, 0, 50), (0, 255, 0), (255, 0, 0)]  # [BLUE, GREEN, RED]
-    print("  Val       R    G    B")
-    print("[")
-    for i in range(steps + 1):
-        val = minval + i * delta
-        r, g, b = convert_to_rgb(minval, maxval, val, colors)
-        print("({:3d}, {:3d}, {:3d}),".format(r, g, b), end="")
-    print("]")
